@@ -13,7 +13,7 @@
     "reskindex": "./reskindex.js"
   },
   "scripts": {
-    "reskindex": "reskindex -h header",
+    "reskindex": "./reskindex.js -h header",
     "build": "babel src -d lib --source-maps --stage 1",
     "start": "babel src -w -d lib --source-maps --stage 1",
     "lint": "eslint src/",
@@ -24,12 +24,13 @@
     "test-multi": "karma start $KARMAFLAGS --single-run=false"
   },
   "dependencies": {
+    "browser-request": "^0.3.3",
     "classnames": "^2.1.2",
     "draft-js": "^0.7.0",
     "draft-js-export-html": "^0.2.2",
     "draft-js-export-markdown": "^0.2.0",
     "draft-js-import-markdown": "^0.1.6",
-    "emojione": "^2.2.2",
+    "emojione": "2.2.3",
     "favico.js": "^0.3.10",
     "filesize": "^3.1.2",
     "flux": "^2.0.3",
@@ -42,10 +43,10 @@
     "matrix-js-sdk": "matrix-org/matrix-js-sdk#develop",
     "optimist": "^0.6.1",
     "q": "^1.4.1",
-    "react": "^15.0.1",
-    "react-addons-css-transition-group": "^15.1.0",
-    "react-dom": "^15.0.1",
-    "react-gemini-scrollbar": "matrix-org/react-gemini-scrollbar#c3d942e",
+    "react": "^15.2.1",
+    "react-addons-css-transition-group": "^15.2.1",
+    "react-dom": "^15.2.1",
+    "react-gemini-scrollbar": "matrix-org/react-gemini-scrollbar#5e97aef",
     "sanitize-html": "^1.11.1",
     "velocity-vector": "vector-im/velocity#059e3b2",
     "whatwg-fetch": "^1.0.0"
