@@ -17,10 +17,12 @@ limitations under the License.
 var q = require("q");
 var React = require('react');
 var MatrixClientPeg = require('../../../MatrixClientPeg');
+var SdkConfig = require('../../../SdkConfig');
 var sdk = require('../../../index');
 var Modal = require('../../../Modal');
 var ObjectUtils = require("../../../ObjectUtils");
 var dis = require("../../../dispatcher");
+var ScalarAuthClient = require("../../../ScalarAuthClient");
 var UserSettingsStore = require('../../../UserSettingsStore');
 
 // parse a string as an integer; if the input is undefined, or cannot be parsed
@@ -65,7 +67,16 @@ module.exports = React.createClass({
             tags_changed: false,
             tags: tags,
             areNotifsMuted: areNotifsMuted,
-            isRoomPublished: false, // loaded async in componentWillMount
+            // isRoomPublished is loaded async in componentWillMount so when the component
+            // inits, the saved value will always be undefined, however getInitialState()
+            // is also called from the saving code so we must return the correct value here
+            // if we have it (although this could race if the user saves before we load whether
+            // the room is published or not).
+            // Default to false if it's undefined, otherwise react complains about changing
+            // components from uncontrolled to controlled
+            isRoomPublished: this._originalIsRoomPublished || false,
+            scalar_token: null,
+            scalar_error: null,
         };
     },
 
@@ -77,6 +88,12 @@ module.exports = React.createClass({
             this._originalIsRoomPublished = result.visibility === "public";
         }, (err) => {
             console.error("Failed to get room visibility: " + err);
+        });
+
+        this.getScalarToken().done((token) => {
+            this.setState({scalar_token: token});
+        }, (err) => {
+            this.setState({scalar_error: err});
         });
 
         dis.dispatch({
@@ -211,10 +228,13 @@ module.exports = React.createClass({
         // color scheme
         promises.push(this.saveColor());
 
+        // url preview settings
+        promises.push(this.saveUrlPreviewSettings());
+
         // encryption
         promises.push(this.saveEncryption());
 
-        console.log("Performing %s operations", promises.length);
+        console.log("Performing %s operations: %s", promises.length, JSON.stringify(promises));
         return q.allSettled(promises);
     },
 
@@ -226,6 +246,11 @@ module.exports = React.createClass({
     saveColor: function() {
         if (!this.refs.color_settings) { return q(); }
         return this.refs.color_settings.saveSettings();
+    },
+
+    saveUrlPreviewSettings: function() {
+        if (!this.refs.url_preview_settings) { return q(); }
+        return this.refs.url_preview_settings.saveSettings();
     },
 
     saveEncryption: function () {
@@ -383,6 +408,37 @@ module.exports = React.createClass({
                 roomState.mayClientSendStateEvent("m.room.guest_access", cli))
     },
 
+    getScalarInterfaceUrl: function() {
+        var url = SdkConfig.get().integrations_ui_url;
+        url += "?scalar_token=" + encodeURIComponent(this.state.scalar_token);
+        url += "&room_id=" + encodeURIComponent(this.props.room.roomId);
+        return url;
+    },
+
+    getScalarToken() {
+        var tok = window.localStorage.getItem("mx_scalar_token");
+        if (tok) return q(tok);
+
+        // No saved token, so do the dance to get one. First, we
+        // need an openid bearer token from the HS.
+        return MatrixClientPeg.get().getOpenIdToken().then((token_object) => {
+            // Now we can send that to scalar and exchange it for a scalar token
+            var scalar_auth_client = new ScalarAuthClient();
+            return scalar_auth_client.getScalarToken(token_object);
+        }).then((token_object) => {
+            window.localStorage.setItem("mx_scalar_token", token_object);
+            return token_object;
+        });
+    },
+
+    onManageIntegrations(ev) {
+        ev.preventDefault();
+        var IntegrationsManager = sdk.getComponent("views.settings.IntegrationsManager");
+        Modal.createDialog(IntegrationsManager, {
+            src: this.state.scalar_token ? this.getScalarInterfaceUrl() : null
+        }, "");
+    },
+
     _renderEncryptionSection: function() {
         if (!UserSettingsStore.isFeatureEnabled("e2e_encryption")) {
             return null;
@@ -415,15 +471,16 @@ module.exports = React.createClass({
         );
     },
 
-
     render: function() {
         // TODO: go through greying out things you don't have permission to change
         // (or turning them into informative stuff)
 
         var AliasSettings = sdk.getComponent("room_settings.AliasSettings");
         var ColorSettings = sdk.getComponent("room_settings.ColorSettings");
+        var UrlPreviewSettings = sdk.getComponent("room_settings.UrlPreviewSettings");
         var EditableText = sdk.getComponent('elements.EditableText');
         var PowerSelector = sdk.getComponent('elements.PowerSelector');
+        var Loader = sdk.getComponent("elements.Spinner")
 
         var cli = MatrixClientPeg.get();
         var roomState = this.props.room.currentState;
@@ -561,6 +618,23 @@ module.exports = React.createClass({
                 </div>
         }
 
+        var integrations_section;
+        if (UserSettingsStore.isFeatureEnabled("integration_management")) {
+            if (this.state.scalar_token) {
+                integrations_section = (
+                    <div className="mx_RoomSettings_settings">
+                        <a href="#" onClick={ this.onManageIntegrations }>Manage integrations</a>
+                    </div>
+                );
+            } else if (this.state.scalar_error) {
+                integrations_section = <div className="error">
+                    Unable to contact integrations server
+                </div>;
+            } else {
+                integrations_section = <Loader />;
+            }
+        }
+
         return (
             <div className="mx_RoomSettings">
 
@@ -571,7 +645,7 @@ module.exports = React.createClass({
                         <input type="checkbox" disabled={ cli.isGuest() }
                                onChange={this._onToggle.bind(this, "areNotifsMuted", true, false)}
                                defaultChecked={this.state.areNotifsMuted}/>
-                        Mute notifications for this room
+                        'Mention only' notifications for this room
                     </label>
                     <div className="mx_RoomSettings_settings">
                         <h3>Who can access this room?</h3>
@@ -645,6 +719,9 @@ module.exports = React.createClass({
                     <ColorSettings ref="color_settings" room={this.props.room} />
                 </div>
 
+                <h3>Integrations</h3>
+                { integrations_section }
+
                 <a id="addresses"/>
 
                 <AliasSettings ref="alias_settings"
@@ -653,6 +730,8 @@ module.exports = React.createClass({
                     canSetAliases={ roomState.mayClientSendStateEvent("m.room.aliases", cli) }
                     canonicalAliasEvent={this.props.room.currentState.getStateEvents('m.room.canonical_alias', '')}
                     aliasEvents={this.props.room.currentState.getStateEvents('m.room.aliases')} />
+
+                <UrlPreviewSettings ref="url_preview_settings" room={this.props.room} />
 
                 <h3>Permissions</h3>
                 <div className="mx_RoomSettings_powerLevels mx_RoomSettings_settings">
@@ -707,7 +786,6 @@ module.exports = React.createClass({
                 <div className="mx_RoomSettings_settings">
                     This room's internal ID is <code>{ this.props.room.roomId }</code>
                 </div>
-
             </div>
         );
     }
